Refactored Vehicle Class to use Rate Object which may vary based on Vehicle attributes.
Introduces a check to determine if Vehicle has AC or NON AC 
so as to apply the neccessary rates or discounts required.

Vehicle is Modelled as at the top of the Inheritance Hierarchy
so exposes some protected state for subclasses

Introduced a polymorphic method which may be overriden by 
Vehicle subclasses. Provided a default implementation based on business policies.

Coded to business rules using domain driven approach that 
hides the implementation of rates calculation here and allows Rate 
Object control and deal with rates Application but for the additional charge if limitIsExceeded since its required for adding up 
the total rate within the calculateTotalExpenseForTrip method.
.

Created a Bus Vehicle Sub-class to capture discount rate difference
and override the implementation of Vehicle

Set up a test case which requires I create a few more Vehicle sub classes
to test it properly Since Vehicle can not be instantiated.